requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
numpy>=1.21.0
sounddevice>=0.4.4
opencv-python>=4.5.3
openai>=0.27.0
websockets>=10.0
python-dotenv>=0.19.0
requests>=2.26.0
asyncio>=3.4.3
aiohttp==3.8.1
numpy==1.21.5
opencv-python==4.5.5.64
sounddevice==0.4.4
soundfile==0.10.3.post1
pyttsx3==2.90
openai==1.3.5
fastapi==0.68.1
python-multipart==0.0.5
uvicorn==0.15.0
python-dotenv==0.19.2
httpx==0.25.2
PyQt5>=5.15.0
qasync>=0.23.0


